Finish updating slots_generated and adjust types for slot IDs

+++ pyo3-ffi/examples/string-sum/src/lib.rs
@@ -60,12 +60,12 @@ static mut SLOTS: [PySlot; SLOTS_LEN] = [
 static mut SLOTS: [PyModuleDef_Slot; SLOTS_LEN] = [
     #[cfg(Py_3_12)]
     PyModuleDef_Slot {
-        slot: Py_mod_multiple_interpreters as c_int,
+        slot: Py_mod_multiple_interpreters,
         value: Py_MOD_PER_INTERPRETER_GIL_SUPPORTED,
     },
     #[cfg(Py_GIL_DISABLED)]
     PyModuleDef_Slot {
-        slot: Py_mod_gil as c_int,
+        slot: Py_mod_gil,
         value: Py_MOD_GIL_NOT_USED,
     },
     PyModuleDef_Slot {
